Fix vxcall

+++ NoVmp/main.cpp
@@ -240,11 +240,11 @@ int main( int argc, const char** argv )
 
 		// Apply optimizations.
 		//
-		size_t ins = rtn->num_instructions();
-		size_t blks = rtn->num_blocks();
+		int64_t ins = (int64_t)rtn->num_instructions();
+		int64_t blks = (int64_t)rtn->num_blocks();
 		vtil::optimizer::apply_all_profiled( rtn );
-		size_t oins = rtn->num_instructions();
-		size_t oblks = rtn->num_blocks();
+		int64_t oins = (int64_t)rtn->num_instructions();
+		int64_t oblks = (int64_t)rtn->num_blocks();
 
 		// Write routine and optimization information.
 		//
